requirements_win.txt
packaging
ninja
diffusers
transformers
tokenizers
accelerate
safetensors
opencv-python
numpy
imageio
imageio-ffmpeg
einops
loguru
qtorch
ftfy
easydict
gradio
aiohttp
pydantic
